py/pyproject.toml
[tool.poetry]
name = "dfo"
version = "0.1.0"
description = "Dispersive Flies Optimisation (DFO) algorithm implementation."
authors = ["Prashant Aparajeya <p.aparajeya@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.4"
aioredis = "^2.0.1"
bcrypt = "^4.1.2"
fastapi="^0.110.1"
gunicorn="^21.2.0"
motor = "^3.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytz = "^2024.1"
uvicorn= {extras = ["bcrypt"], version = "^0.29.0"}
websockets = "^12.0"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
